before switch to jar

+++ src/main/kotlin/event/ParkEvent.kt
@@ -1,5 +1,6 @@
 package event
 
+import kotlinx.datetime.Instant
 import org.kalasim.Event
 import org.kalasim.TickTime
 
@@ -7,5 +8,6 @@ class ParkEvent(
     val powerPlantId: Int,
     val powerPlantName: String,
     //val parkType: String,   TODO kell???
-    val producedPower: Float,
-    time: TickTime): Event(time)
+    val producedPower: Double,
+    time: TickTime
+): Event(time)
